feat(ANSH-99): add logout for photographer settings section
