+++ app/user/(routes)/photographer/[userId]/(routes)/settings/components/earnings.tsx
