+++ app/user/(routes)/photographer/[userId]/(routes)/settings/components/faq.tsx
