+++ app/user/(routes)/photographer/[userId]/(routes)/settings/components/paymentColumns.tsx
