+++ app/user/layout.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { Button } from "@/components/ui/button";
-import { signOut } from "next-auth/react";
+// import { signOut } from "next-auth/react";
 
 type UserLayoutProps = {
     children: React.ReactNode;
@@ -10,11 +10,11 @@ type UserLayoutProps = {
 const UserLayout = (props: UserLayoutProps) => {
     return (
         <>
-            <div>
+            {/* <div>
                 <Button onClick={() => signOut({
                     callbackUrl: `${window.location.origin}/sign-in`
                 })}>Sign Out</Button>
-            </div>
+            </div> */}
             {props.children}
         </>
     );
